New Custom Color DIalog
lets face it, the default QColorDialog is ugly. so after months of working i came up with this new design. unfortunately it require additional dependency which is pyinput for the color picker. but i decided to try except it in case anyone not notice. i plan to add this into separate repository just so another project that require QColorDialog won't suffer like i did.